Refactor dataloader getter in UnitTest
