+++ test/test_engine.py
@@ -1,5 +1,4 @@
 # Copyright (c) 2021, Zhiqiang Wang. All Rights Reserved.
-from pathlib import Path
 import unittest
 
 import torch
@@ -10,16 +9,12 @@ from torchvision.io import read_image
 import pytorch_lightning as pl
 
 from yolort.data import DetectionDataModule
-from yolort.data.coco import CocoDetection
-from yolort.data.transforms import collate_fn, default_train_transforms
 
 from yolort.models.yolo import yolov5_darknet_pan_s_r31
 from yolort.models.transform import nested_tensor_from_tensor_list
 from yolort.models import yolov5s
 
-from yolort.utils import prepare_coco128
-
-from .dataset_utils import DummyCOCODetectionDataset
+from yolort.utils.dataset_utils import DummyCOCODetectionDataset, get_data_loader
 
 from typing import Dict
 
@@ -65,26 +60,9 @@ class EngineTester(unittest.TestCase):
         # Define the device
         device = torch.device('cpu')
 
-        # Prepare the datasets for training
-        # Acquire the images and labels from the coco128 dataset
-        data_path = Path('data-bin')
-        coco128_dirname = 'coco128'
-        coco128_path = data_path / coco128_dirname
-        image_root = coco128_path / 'images' / 'train2017'
-        annotation_file = coco128_path / 'annotations' / 'instances_train2017.json'
-
-        if not annotation_file.is_file():
-            prepare_coco128(data_path, dirname=coco128_dirname)
-
-        batch_size = 4
-
-        dataset = CocoDetection(image_root, annotation_file, default_train_transforms())
-        sampler = torch.utils.data.RandomSampler(dataset)
-        batch_sampler = torch.utils.data.BatchSampler(sampler, batch_size, drop_last=True)
-        data_loader = torch.utils.data.DataLoader(
-            dataset, batch_sampler=batch_sampler, collate_fn=collate_fn, num_workers=0)
+        train_dataloader = get_data_loader(mode='train')
         # Sample a pair of images/targets
-        images, targets = next(iter(data_loader))
+        images, targets = next(iter(train_dataloader))
         images = [img.to(device) for img in images]
         targets = [{k: v.to(device) for k, v in t.items()} for t in targets]
 
@@ -109,19 +87,17 @@ class EngineTester(unittest.TestCase):
         trainer = pl.Trainer(max_epochs=1)
         trainer.fit(model, data_module)
 
-    def test_test_dataloaders(self):
-        # Config dataset
-        num_samples = 128
-        batch_size = 4
-        # Setup the DataModule
-        train_dataset = DummyCOCODetectionDataset(num_samples=num_samples)
-        data_module = DetectionDataModule(train_dataset, batch_size=batch_size)
+    @unittest.skip("Currently it isn't well implemented")
+    def test_test_with_dataloader(self):
+        # Get dataloader to test
+        val_dataloader = get_data_loader(mode='val')
+
         # Load model
         model = yolov5s(pretrained=True)
         model.eval()
         # Trainer
         trainer = pl.Trainer(max_epochs=1)
-        trainer.test(model, test_dataloaders=data_module.val_dataloader(batch_size=batch_size))
+        trainer.test(model, test_dataloaders=val_dataloader)
 
     def test_predict_with_vanilla_model(self):
         # Set image inputs
