+++ yolort/utils/__init__.py
@@ -1,4 +1,3 @@
 from .flash_utils import get_callable_dict
 from .image_utils import cv2_imshow, get_image_from_url, read_image_to_tensor
 from .update_module_state import update_module_state_from_ultralytics
-from .file_utils import prepare_coco128
